requirements.txt
requests
python-docx
pytesseract
Pillow